fix(components/molecule/pagination): refactor
