+++ components/molecule/pagination/src/settings.js
@@ -1,0 +1,8 @@
+export const BASE_CLASS = 'sui-MoleculePagination'
+export const PAGE_NUMBER_HOLDER = '%{pageNumber}'
+export const DIVIDER = '···'
+
+export const defaultCreateUrl = ({pageNumber, urlPattern}) =>
+urlPattern.replace(PAGE_NUMBER_HOLDER, pageNumber)
+
+export const noop = () => {}
